test: extract createMockVersion into a shared helper
