+++ test/cooldown.test.ts
@@ -11,8 +11,8 @@ import { npmApi } from '../src/package-managers/npm'
 import { pnpmApi } from '../src/package-managers/pnpm'
 import { yarnApi } from '../src/package-managers/yarn'
 import type { PackageFile } from '../src/types/PackageFile'
-import type { Packument } from '../src/types/Packument'
 import chaiSetup from './helpers/chaiSetup'
+import createMockVersion from './helpers/createMockVersion'
 import { silenceProgressBar } from './helpers/silenceProgressBar'
 import stubVersions from './helpers/stubVersions'
 
@@ -20,31 +20,6 @@ chaiSetup()
 
 const DAY = 24 * 60 * 60 * 1000
 const NOW = Date.now()
-
-interface CreateMockParams {
-  name: string
-  versions: Record<string, string>
-  distTags?: Record<string, string>
-}
-
-/**
- * Creates a mock package version object for testing purposes.
- *
- * @param params - The parameters for creating the mock version.
- * @param params.name - The name of the package.
- * @param params.versions - An object mapping version strings to their corresponding release dates.
- * @param params.distTags - An object representing distribution tags for the package.
- * @returns An object representing mocked package versions, including name, versions, time, and distTags.
- */
-const createMockVersion = ({ name, versions, distTags }: CreateMockParams): Partial<Packument> => {
-  return {
-    name,
-    version: Object.keys(versions)[0],
-    versions: Object.fromEntries(Object.entries(versions).map(([version]) => [version, { version } as Packument])),
-    time: Object.fromEntries(Object.entries(versions).map(([version, date]) => [version, date])),
-    'dist-tags': distTags,
-  }
-}
 
 /**
  * Processes log arguments by removing ANSI escape codes, collapsing all
